feat: migrates styleguide to app directory
