+++ app/styleguide/icons/head.tsx
@@ -1,0 +1,11 @@
+import { Fragment } from 'react'
+
+export default function Head() {
+  return (
+    <Fragment>
+      <meta key="robots" name="robots" content="noindex,follow" />
+      <meta key="googlebot" name="googlebot" content="noindex,follow" />
+      <title>Icons</title>
+    </Fragment>
+  )
+}
