+++ components/Button/index.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import Link from 'next/link'
 import React, { useCallback } from 'react'
 import { Reference } from '../CMSLink'
